работа без таблицы сотрудников
git-svn-id: https://share.curs.ru/svn/grains/celestabasic/trunk/common.sys@1138 bf7a05a2-0d8f-8241-91a8-991a96a8b1b0

+++ score/security/setForeignKeys.py
@@ -18,34 +18,38 @@ def setForeignKeys(context):
         # Если пользователи приходят из mellophone, поле password в таблице logins становится nullable.
         password_field = login_table.getColumn("password") #получаем объект поля password таблицы logins
         password_field.setNullableAndDefault(settings.isUseAuthServer(), None) #Разрешаем значения null для него
-    employees_grain=score.getGrain(settings.getEmployeesParam("employeesGrain")) #получаем объект гранулы, которая содержит таблицу сотрудников
-    employees_table=employees_grain.getTable(settings.getEmployeesParam("employeesTable")) #получаем объект таблицы сотрудников
-    employees_id = employees_table.getColumn(settings.getEmployeesParam("employeesId")) #получаем объект колонки первичного ключа таблицы сотрудников
-    if isinstance(employees_id, IntegerColumn):
-        # если первичный ключ таблицы сотрудников целочисленный, удаляем колонку "employeeId" таблицы subjects 
-        # и создаём новую целочисленного типа.
-        # сделано так потому что по умолчанию поле "employeeId" таблицы subjects текстовое.
-        subject_table.getColumn("employeeId").delete()
-        IntegerColumn(subject_table, "employeeId")
-    else:
-        # если первичный ключ таблицы сотрудников текстовый, устанавливаем длину поля "employeeId" таблицы subjects
-        # в соответствии с ним
-        employees_id_length = employees_id.getLength()
-        subject_table.getColumn("employeeId").setLength(unicode(employees_id_length))
+    if settings.isEmployees():
+        employees_grain=score.getGrain(settings.getEmployeesParam("employeesGrain")) #получаем объект гранулы, которая содержит таблицу сотрудников
+        employees_table=employees_grain.getTable(settings.getEmployeesParam("employeesTable")) #получаем объект таблицы сотрудников
+        employees_id = employees_table.getColumn(settings.getEmployeesParam("employeesId")) #получаем объект колонки первичного ключа таблицы сотрудников
+        if isinstance(employees_id, IntegerColumn):
+            # если первичный ключ таблицы сотрудников целочисленный, удаляем колонку "employeeId" таблицы subjects 
+            # и создаём новую целочисленного типа.
+            # сделано так потому что по умолчанию поле "employeeId" таблицы subjects текстовое.
+            subject_table.getColumn("employeeId").delete()
+            IntegerColumn(subject_table, "employeeId")
+        else:
+            # если первичный ключ таблицы сотрудников текстовый, устанавливаем длину поля "employeeId" таблицы subjects
+            # в соответствии с ним
+            employees_id_length = employees_id.getLength()
+            subject_table.getColumn("employeeId").setLength(unicode(employees_id_length))
+        
+        subjects_keys = subject_table.getForeignKeys() #получаем список объектов-внешних ключей таблицы subjects
+        for subjects_key in subjects_keys:
+            #удаляем все внешние ключи таблицы subjects
+            subjects_key.delete()
+        subjects_key = ForeignKey(subject_table, employees_table, ["employeeId"]) #создаём внешний ключ subjects(employeeId)-<employees>(<employeeId>)
+        
     logins_keys = login_table.getForeignKeys() #получаем список объектов-внешних ключей таблицы logins
     for logins_key in logins_keys:
         #удаляем все внешние ключи таблицы logins
         logins_key.delete()                                
     logins_key = ForeignKey(login_table, subject_table, ["subjectId"]) #создаём внешний ключ logins(subjectId)-subjects(sid)
-    subjects_keys = subject_table.getForeignKeys() #получаем список объектов-внешних ключей таблицы subjects
-    for subjects_key in subjects_keys:
-        #удаляем все внешние ключи таблицы subjects
-        subjects_key.delete()
-    subjects_key = ForeignKey(subject_table, employees_table, ["employeeId"]) #создаём внешний ключ subjects(employeeId)-<employees>(<employeeId>)
     if settings.loginIsSubject(): #устанавливаем правила удаления 
         # Если logins тождественны subjects:
         logins_key.setDeleteRule(FKRule.CASCADE) # каскадное удаление на ключ logins-subjecs
-        subjects_key.setDeleteRule(FKRule.SET_NULL) # установка null на ключ subjects-<employees>
+        if settings.isEmployees():
+            subjects_key.setDeleteRule(FKRule.SET_NULL) # установка null на ключ subjects-<employees>
     else:
         logins_key.setDeleteRule(FKRule.SET_NULL) # установка null на ключ logins-subjecs
     score.save() #сохраняем изменения
